app/backend/requirements.txt
flask
flask-cors

python-dotenv
azure-storage-blob
requests
beautifulsoup4
tiktoken

crawlee
playwright

azure-search-documents>=11.4.0
numpy
azure-cosmos

openai
tenacity

gremlinpython
